+++ include/beast/http/string_body.hpp
@@ -47,8 +47,6 @@ struct string_body
         value_type const& body_;
 
     public:
-        using is_deferred = std::false_type;
-
         using const_buffers_type =
             boost::asio::const_buffers_1;